chore: organize component directory structure
