+++ client/src/components/atoms/headerAtoms/FridgeSelector/FridgeSelector.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import FridgeSelector from "@/components/molecules/Header/headerAtoms/FridgeSelector/FridgeSelector";
+import FridgeSelector from "@/components/atoms/headerAtoms/FridgeSelector/FridgeSelector";
 import type { SelectedView } from "@/components/molecules/Header/Header";
 
 const meta: Meta<typeof FridgeSelector> = {
